@@ -1,14 +1,13 @@
 package esz.dev;
 
-import esz.dev.argparse.ArgParseException;
 import esz.dev.argparse.ArgumentParser;
 import esz.dev.delaunay.Delaunay;
 import esz.dev.delaunay.DelaunayException;
 
 import java.io.IOException;
 
-public class Main {
 
+public class Main {
     public static void main(String[] args) {
 
         nu.pattern.OpenCV.loadShared();
@@ -18,7 +17,7 @@ public class Main {
             argumentParser.process();
             Delaunay delaunay = new Delaunay(argumentParser.process());
             delaunay.generate();
-        } catch (ArgParseException e) {
+        } catch (IllegalArgumentException e) {
             System.out.println("Error: " + e.getMessage());
             ArgumentParser.showHelp();
         } catch (DelaunayException | IOException e) {
