@@ -4,21 +4,21 @@ import org.opencv.core.Point;
 
 public class Triangle {
 
-    private Edge ab;
-    private Edge bc;
-    private Edge ca;
-    private Circle circumCircle;
+    private final Edge ab;
+    private final Edge bc;
+    private final Edge ca;
+    private final Circle circumCircle;
 
     public Triangle(Point a, Point b, Point c) {
         this.ab = new Edge(a, b);
         this.bc = new Edge(b, c);
         this.ca = new Edge(c, a);
 
-        calculateCircumCircle(a, b, c);
+        this.circumCircle = calculateCircumCircle(a, b, c);
     }
 
     // https://en.wikipedia.org/wiki/Circumscribed_circle#Triangles
-    private void calculateCircumCircle(Point a, Point b, Point c) {
+    private static Circle calculateCircumCircle(Point a, Point b, Point c) {
         double d = 2 * (a.x * (b.y - c.y) + b.x * (c.y - a.y) + c.x * (a.y - b.y));
         double centerX = 1.0 / d * ((a.x * a.x + a.y * a.y) * (b.y - c.y) + (b.x * b.x + b.y * b.y) * (c.y - a.y) + (c.x * c.x + c.y * c.y) * (a.y - b.y));
         double centerY = 1.0 / d * ((a.x * a.x + a.y * a.y) * (c.x - b.x) + (b.x * b.x + b.y * b.y) * (a.x - c.x) + (c.x * c.x + c.y * c.y) * (b.x - a.x));
@@ -30,7 +30,7 @@ public class Triangle {
         double centerYSign = 1.0 / dSign * (bSign.x * (cSign.x * cSign.x + cSign.y * cSign.y) - cSign.x * (bSign.x * bSign.x + bSign.y * bSign.y));
 
         double r = Math.sqrt(centerXSign * centerXSign + centerYSign * centerYSign);
-        this.circumCircle = new Circle(new Point(centerX, centerY), r);
+        return new Circle(new Point(centerX, centerY), r);
     }
 
     public Circle getCircumCircle() {
