@@ -17,28 +17,24 @@ import org.opencv.core.Size;
 import java.awt.*;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
+import java.util.function.BiConsumer;
 
 public class VectorImageCreator extends ImageCreator {
 
-    private Graphics2D vg2d = new VectorGraphics2D();
+    private final Graphics2D vg2d = new VectorGraphics2D();
 
-    @FunctionalInterface
-    private interface Style {
-        void draw(int[] xVec, int[] yVec);
-    }
+    private final BiConsumer<int[], int[]> drawFunction;
 
-    private Style style;
-
-    VectorImageCreator(ArrayList<Triangle> triangles, Mat originalImage, FillColorInterface fillColor,
-                              String outputPath, boolean wireFrame) {
+    VectorImageCreator(List<Triangle> triangles, Mat originalImage, FillColorInterface fillColor,
+                       String outputPath, boolean wireFrame) {
         super(triangles, originalImage, fillColor, outputPath, wireFrame);
 
         if (wireFrame) {
-            style = (int[] xVec, int[] yVec) -> vg2d.fillPolygon(xVec, yVec, 3);
+            drawFunction = (int[] xVec, int[] yVec) -> vg2d.fillPolygon(xVec, yVec, 3);
         } else {
-            style = (int[] xVec, int[] yVec) -> {
+            drawFunction = (int[] xVec, int[] yVec) -> {
                 vg2d.fillPolygon(xVec, yVec, 3);
                 vg2d.drawPolygon(xVec, yVec, 3);
             };
@@ -47,11 +43,11 @@ public class VectorImageCreator extends ImageCreator {
 
     @Override
     public void createTriangle(Point[] vertices) {
-        int[] xVec = {(int)Math.ceil(vertices[0].x), (int)Math.ceil(vertices[1].x), (int)Math.ceil(vertices[2].x)};
-        int[] yVec = {(int)Math.ceil(vertices[0].y), (int)Math.ceil(vertices[1].y), (int)Math.ceil(vertices[2].y)};
+        int[] xVec = {(int) Math.ceil(vertices[0].x), (int) Math.ceil(vertices[1].x), (int) Math.ceil(vertices[2].x)};
+        int[] yVec = {(int) Math.ceil(vertices[0].y), (int) Math.ceil(vertices[1].y), (int) Math.ceil(vertices[2].y)};
         Scalar scalar = fillColor.getFillColor(originalImage, vertices);
         vg2d.setColor(scalarToColor(scalar));
-        style.draw(xVec, yVec);
+        drawFunction.accept(xVec, yVec);
     }
 
     @Override
@@ -76,9 +72,9 @@ public class VectorImageCreator extends ImageCreator {
     private Color scalarToColor(Scalar scalar) {
         double[] colorVec = scalar.val;
         if (colorVec.length == 3) {
-            return new Color((int)scalar.val[2], (int)scalar.val[1], (int)scalar.val[0]);
+            return new Color((int) scalar.val[2], (int) scalar.val[1], (int) scalar.val[0]);
         } else {
-            return new Color((int)scalar.val[0], (int)scalar.val[0], (int)scalar.val[0]);
+            return new Color((int) scalar.val[0], (int) scalar.val[0], (int) scalar.val[0]);
         }
     }
 }
